+++ abstraction/olcs/index.js
@@ -1,1 +1,1 @@
-export {default as map} from "./map";
+export {default as olcsMap} from "./olcsMap";