+++ src/index.js
@@ -2,6 +2,7 @@ import {createMapView} from "./mapView";
 import * as crs from "./crs";
 import * as rawLayerList from "./rawLayerList";
 import * as wms from "./layer/wms";
+import * as wfs from "./layer/wfs";
 import * as geojson from "./layer/geojson";
 import * as vectorBase from "./layer/vectorBase";
 import * as layerLib from "./layer/lib";
@@ -11,9 +12,10 @@ import setBackgroundImage from "./lib/setBackgroundImage";
 export {
     createMapView,
     wms,
-    vectorBase,
+    wfs,
     geojson,
     layerLib,
+    vectorBase,
     setBackgroundImage,
     setGazetteerUrl,
     rawLayerList,